Moved log logic from chat.py to a separate file -> utility.log_event.py : duplicate code put in a separate function : removed unnecessary hard coded value
