+++ src/backend/endpoints/log_event.py
@@ -40,55 +40,3 @@ async def log_event(payload: LogEventPayload, user=Depends(get_current_user)):
     )
 
     langfuse.flush()
-
-
-active_trace_ids: dict[str, str] = {}
-
-def start_session(user_id: str, sid: str) -> str:
-    langfuse = get_client()
-    trace_id = langfuse.create_trace_id()
-    active_trace_ids[sid] = trace_id
-
-    with langfuse.start_as_current_observation(
-        as_type="span",
-        name="session_start",
-        trace_context={"trace_id": trace_id, "parent_span_id": "0123456789abcdef"},
-    ) as observation:
-        observation.update(
-            input={"event": "connect", "user_id": user_id, "sid": sid},
-            output={"status": "connected", "message": "User connected via Socket.IO"},
-        )
-        observation.update_trace(
-            user_id=user_id,
-            session_id=sid,
-            metadata={"message": "User connected via Socket.IO"},
-        )
-
-    langfuse.flush()
-    return trace_id
-
-
-def end_session(user_id: str | None, sid: str):
-    trace_id = active_trace_ids.pop(sid, None)
-    if trace_id:
-        langfuse = get_client()
-
-        with langfuse.start_as_current_observation(
-            as_type="span",
-            name="session_end",
-            trace_context={"trace_id": trace_id, "parent_span_id": "0123456789abcdef"},
-        ) as observation:
-            observation.update(
-                input={"event": "disconnect", "user_id": user_id, "sid": sid},
-                output={
-                    "status": "disconnected",
-                    "message": "User disconnected / socket closed",
-                },
-            )
-            observation.update_trace(
-                user_id=user_id,
-                session_id=sid,
-                metadata={"message": "User disconnected / socket closed"},
-            )
-
-        langfuse.flush()
